Move intercom notification to a separate subproject
