+++ app-backend/batch/src/main/resources/application.conf
@@ -113,11 +113,3 @@ s3 {
   region = ${?AWS_REGION}
   dataBucket = ${?DATA_BUCKET}
 }
-
-intercom {
-  adminId = ${?INTERCOM_ADMIN_ID}
-
-  token = ${?INTERCOM_TOKEN}
-
-  groundworkUrlBase = ${?GROUNDWORK_URL_BASE}
-}
