+++ app-backend/notification/src/main/resources/reference.conf
@@ -1,0 +1,7 @@
+intercom {
+  adminId = ${?INTERCOM_ADMIN_ID}
+
+  token = ${?INTERCOM_TOKEN}
+
+  groundworkUrlBase = ${?GROUNDWORK_URL_BASE}
+}
